Added the logic to allow a user to add an amount to warrant their product
Originally, I planned to have the default value set to the price of the product. However, when the page was rendered, this value wasn't set. So to have it show up, I would have to make player.round a dependency, which meant tinkering with the core file, which I couldn't do. Therefore i just set it to a default text value.

Also added to the function handleWarrantAmount which now parses the input and sets the warrantAmount based on it.

Also Made warrantAmount a state hook so that the ui dynamically updates.